[Refactoring] Move noise-related functions > utils
* Move all noise-related functions into a separate utils script

* Add & Update documentation
